added sqlite3 state backend
Some users may want to organize their photos post downloading which is not possible given the only prior behavior of looking for the photo by name in the download directory.

This adds the ability to back the already downloaded files state with sqlite3 so we can manipulate the files afterwards without redownloading.

It also uses a transactional model so any partially downlaoded files will be re-downloaded on the next attempt